# src/main/resources/swagger.yml
openapi: 3.0.0
info:
  title: AI Summary Transactions API
  description: It allows you to summarize transactions using AI.
  version: 1.0.0
  license:
    name: Apache-2.0
    url: "https://www.apache.org/licenses/LICENSE-2.0.html"
servers:
  - url: http://localhost:8080
    description: Local server
paths:
...
  /ai/transactions/summary:
    post:
      tags:
        - AI transactions
      summary: AI process transaction by generated query
      description: Process a natural language and generate a query to search for transactions using AI
      operationId: processAITransactionSummary
      parameters:
        - $ref: "#/components/parameters/StartDateRequired"
        - $ref: "#/components/parameters/EndDateRequired"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                question:
                  type: string
                  description: Natural language question about transactions
                  example: "Quais foram as transações de hoje?"
              required:
                - question
      responses:
        "200":
          $ref: "#/components/responses/AIResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /ai/transactions/insights/{scenario}:
    post:
      tags:
        - AI transactions
      summary: Get AI insights for transactions by scenario
      description: Generate AI-powered insights for transactions based on a specific scenario within a date range
      operationId: getAITransactionInsights
      parameters:
        - $ref: "#/components/parameters/Scenario"
        - $ref: "#/components/parameters/StartDateRequired"
        - $ref: "#/components/parameters/EndDateRequired"
      responses:
        "200":
          $ref: "#/components/responses/AIResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
components:
  parameters:
...
    StartDateRequired:
      name: startDate
      in: query
      required: true
      description: Start date for filtering transactions
      schema:
        type: string
        format: date
        example: "2025-01-01"
    EndDateRequired:
      name: endDate
      in: query
      required: true
      description: End date for filtering transactions
      schema:
        type: string
        format: date
        example: "2025-01-31"
...
    Scenario:
      name: scenario
      in: path
      required: true
      description: The scenario type for generating insights
      schema:
        type: string
        enum: [Fechada, Aberta, Futura]
        example: "Fechada"
  responses:
    BadRequest:
      description: Bad request - invalid input data
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    AIResponse:
      description: AI processed query successfully
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/AIResultResponse"
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "Error message"
    AIResultResponse:
      type: object
      properties:
        result:
          type: string
          description: AI processed result
          example: "Encontrei 5 transações de hoje: compra no supermercado, pagamento de conta, etc."
...